+++ src/javaharness/java/arcs/crdt/RawCollection.java
@@ -3,7 +3,7 @@ package arcs.crdt;
 import java.util.HashSet;
 import java.util.List;
 
-public class RawCollection<T> extends HashSet<T> implements CRDTConsumerType {
+class RawCollection<T> extends HashSet<T> implements CRDTConsumerType {
   RawCollection() {}
 
   RawCollection(List<T> list) {
