+++ src/javaharness/javatests/arcs/android/impl/BUILD
@@ -10,7 +10,6 @@ java_test(
         "//java/arcs/android/impl:impl",
         "@junit//jar",
         "//java/arcs/api:api-android",
-        "//java/arcs/crdt:crdt-android",
     ],
 )
